refactor: migrate MetaTrader 5 server codebase to new architecture
