+++ pyproject.toml
@@ -20,13 +20,11 @@ classifiers = [
     "Topic :: Office/Business :: Financial :: Investment",
 ]
 dependencies = [
-    "fastmcp>=0.4.1",
-    "httpx>=0.28.1",
-    "mcp[cli]>=1.6.0",
-    "metatrader5>=5.0.4874",
-    "pandas>=2.2.3",
-    "numpy>=1.24.0",
-    "pydantic>=2.0.0",
+    "fastmcp==2.12.4",
+    "metatrader5==5.0.5260",
+    "pandas>=2.3.3",
+    "pydantic>=2.11.9",
+    "python-dotenv>=1.1.1",
 ]
 
 [project.urls]
@@ -34,7 +32,7 @@ dependencies = [
 "Bug Tracker" = "https://github.com/Qoyyuum/mcp-metatrader5-server/issues"
 
 [project.scripts]
-mt5mcp = "mcp_metatrader5_server.cli:main"
+mt5mcp = "mcp_mt5:main"
 
 [tool.setuptools]
 package-dir = {"" = "src"}
