--- src/mcp_metatrader5_server/__init__.py
@@ -1,7 +1,0 @@
-"""
-MCP MetaTrader 5 Server - A Model Context Protocol server for MetaTrader 5.
-
-This package provides tools and resources for AI assistants to interact with MetaTrader 5.
-"""
-
-__version__ = "0.1.0"
