+++ src/mcp_mt5/.env
@@ -1,0 +1,3 @@
+MT5_MCP_HOST=127.0.0.1
+MT5_MCP_PORT=8000
+MT5_PATH="C:\Program Files\RoboForex MT5 Terminal\terminal64.exe"
